move event distributor under send_sms context and add logging
