+++ serverless.yml
@@ -101,7 +101,7 @@ functions:
     timeout: 60  # Give aurora cluster time to cold start
 
   distributeDialogEvents:
-    handler: stopcovid/event_distributor/event_distributor.distribute_dialog_events
+    handler: aws_lambdas/enqueue_sms_batch.handler
     events:
       - stream:
           type: dynamodb
